cart functions connected to backend

+++ FeaturesApp/models.py
@@ -28,7 +28,7 @@ class Product(models.Model):
     is_organic = models.BooleanField(default=False)
     rating = models.FloatField(default=0)
     review_count = models.PositiveIntegerField(default=0)
-    discount = models.FloatField(default=0)
+    discount = models.DecimalField(max_digits=5, decimal_places=2, default=0)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
